pass package filtering through action
