+++ .changes/filter-packages.md
@@ -1,6 +1,7 @@
 ---
 "covector": patch
 "@covector/assemble": minor
+"action": minor
 ---
 
 Some workflows require different actions for different packages. Most of this can be codified into config. However there are cases where you may need to run a command for a dynamic set of packages.
